django-backend/app/requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-environ==0.11.2
djangorestframework==3.15.2
environ==1.0
exceptiongroup==1.2.2
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ollama==0.3.0
packaging==24.1
psycopg2-binary==2.9.9
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
zope.event==5.0
zope.interface==6.4.post2
